pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Ciastelix", email = "tenlegitmati@gmail.com"},
]
dependencies = [
    "dependency-injector>=4.41.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.1",
    "sqlalchemy>=2.0.23",
    "passlib>=1.7.4",
    "uvicorn>=0.24.0.post1",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

